requirements.txt
absl-py==1.0.0
arrow==1.2.1
astunparse==1.6.3
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
cycler==0.11.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
flatbuffers==2.0
gast==0.4.0
google-api-core==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-cloud-pubsub==2.9.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.2
jinja2-time==0.2.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
libcst==0.3.23
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
poyo==0.5.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.4
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
xgboost==1.5.0
zipp==3.6.0
